String to number conversion fix
git-svn-id: svn://cherokee-project.com/cherokee/trunk@5506 5dc97367-97f1-0310-9951-d761b3857238

+++ admin/market/Util.py
@@ -109,7 +109,7 @@ class PriceTag (CTK.Box):
     def __init__ (self, info):
         CTK.Box.__init__ (self, {'class': 'market-app-price'})
 
-        if int(info['amount']):
+        if float(info['amount']):
             self += CTK.Box ({'class': 'currency'}, CTK.RawHTML (info['currency_symbol']))
             self += CTK.Box ({'class': 'amount'},   CTK.RawHTML (info['amount']))
             self += CTK.Box ({'class': 'currency'}, CTK.RawHTML (info['currency']))
